[GTK][AC] Animating layer disappears while running with clutter backend https://bugs.webkit.org/show_bug.cgi?id=110470
Patch by ChangSeok Oh <changseok.oh@collabora.com> on 2013-03-28
Reviewed by Gustavo Noronha Silva.

The reason of this issue is that clutter_actor_remove_child leads to stopping animations
of child actor. ClutterActor's animation could be defined only when it has a parent actor.
So we should avoid the case calling the api as much as we can. At least we don't
need to reset a actor's parent at all if the current parent is same with new one.
If we can't avoid invoking clutter_actor_remove_child for an animating actor, we may apply
a more complicated way to keep the animation. But I haven't faced such a case yet.

Covered by existing animation tests.

* platform/graphics/clutter/GraphicsLayerActor.cpp:
(graphicsLayerActorSetSublayers):
* platform/graphics/clutter/GraphicsLayerClutter.cpp:
(WebCore::GraphicsLayerClutter::recursiveCommitChanges):
(WebCore::GraphicsLayerClutter::updateSublayerList):

Identifier: 131848@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@147121 268f45cc-cd09-0410-ab3c-d52691b4dbfc