[wk2] REGRESSION (r125500): WebFrameLoaderClient's active m_pluginView can be zeroed by PluginDocuments leaving the page cache https://bugs.webkit.org/show_bug.cgi?id=113430 <rdar://problem/12729564>
Reviewed by Brady Eidson.

Ignore redirectDataToPlugin(0), which is only called from PluginDocument::detach(),
and only after r125500. Otherwise, PluginDocuments that are put into the page cache
but leave later can clear a totally different and legitimate PluginView from the
WebFrameLoaderClient, causing data to be lost and PluginDocument loads to stall.

This restores our behavior to the pre-r125500 state, which should be fine, as we
clear WebFrameLoaderClient's m_pluginView when the load fails or finishes. This
matches the changes made in r125500 for every platform other than Chromium and WebKit2.

* WebProcess/WebCoreSupport/WebFrameLoaderClient.cpp:
(WebKit::WebFrameLoaderClient::redirectDataToPlugin):


Identifier: 131877@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@147152 268f45cc-cd09-0410-ab3c-d52691b4dbfc